requirements-image-edit-test.txt
Flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
Pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0
opencv-python>=4.8.0
